+++ index.js
@@ -1,9 +1,0 @@
-// imports
-import observe from './src/observe';
-import style from './src/style';
-
-// polyfill
-export {
-	observe,
-	style
-};