+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to hfill
 
+### 2.0.0 (May 28, 2017)
+
+- Harmonize API with hfill
+- Simplify polyfill and limit usage to `h` element
+
 ### 1.0.0 (January 17, 2017)
 
 - Initial version